Remove temporary files
Process started on 21 Feb 2021 to reconcile while model stopped working well after implementing several changes to align it with 2015 U-net paper is now finished. The main driver was the learning rate scheduler which was placed at a wrong position so the learning rate decreased rapidly and the model stopped learning.